Completely redid build system.
The top level makefile now enteres the sys directory recursively as it did before
but now it invokes the sys/Makefile i created. this makes building faster as the new
makefile only recompiles source that has changed. it is also easier to follow the warnings
as they are color coded and not mixed in between directory changes.